% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mat.R
\name{calculate_complete_loglik}
\alias{calculate_complete_loglik}
\title{Calculate the Complete-Data Log-Likelihood for a Given Model}
\usage{
calculate_complete_loglik(beta_vec, nu_vec, lambda_vec, X, y, model_params)
}
\arguments{
\item{beta_vec, }{nu_vec, lambda_vec The current p-dimensional vectors for the latent variables.}

\item{X, }{y The data.}

\item{model_params}{A list containing the specific hyperparameters (a, b, omega, sigmaSq) for a candidate model.}
}
\value{
A single numeric value representing the complete-data log-likelihood.
}
\description{
This function computes the logarithm of p(y|beta, model_k)xp(beta|lambda, nu, model_k)xp(lambda|model_k)xp(nu|model_k)
which serves as the weight in the model selection step.
}
