% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sigmaSq_omega.R
\name{estimate_omega_mom}
\alias{estimate_omega_mom}
\title{Estimate Omega using Method of Moments given estimate of sigmaSq}
\usage{
estimate_omega_mom(X, y, sigmaSq_hat, epsilon = 1e-06)
}
\arguments{
\item{X}{Numeric matrix. The design matrix of dimension n x p.}

\item{y}{Numeric vector. The response vector of length n.}

\item{sigmaSq_hat}{Numeric scalar. The estimate of error variance (e.g., from ridge or organic lasso).}

\item{epsilon}{Numeric scalar. A small positive value to return if the estimate is non-positive. Default is 1e-6.}
}
\value{
Numeric scalar. The estimated value of omega.
}
\description{
Estimate Omega using Method of Moments given estimate of sigmaSq
}
