% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FullGibbs.R
\name{fullGibbs}
\alias{fullGibbs}
\title{call getsamples.full function}
\usage{
fullGibbs(
  burnin_num,
  output_num,
  X,
  y,
  a0,
  b0,
  beta0,
  nu0,
  xi0 = NULL,
  lambda0,
  omega0,
  w0,
  sigmaSq0,
  omega_hyper,
  tunea = 0.1,
  tuneb = 0.2,
  uppera = 1,
  upperb = 1,
  woodbury = FALSE,
  approx = FALSE,
  diagX = FALSE,
  prshapeo,
  prrateo,
  prshapes,
  prrates
)
}
\description{
call getsamples.full function
}
