% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sigmaSq_omega.R
\name{get_initial_values}
\alias{get_initial_values}
\title{Initialization Wrapper
A wrapper to bundle the sigmaSq and omega initialization.}
\usage{
get_initial_values(X, y, option = "ridge")
}
\arguments{
\item{X}{Numeric matrix.}

\item{y}{Numeric vector.}

\item{option}{Binary variable, either "olasso" for oracle lasso with lam = log(p)/n or "ridge" based on ridge regression with CV.}
}
\value{
A list containing initialized omega and sigmaSq.
}
\description{
Initialization Wrapper
A wrapper to bundle the sigmaSq and omega initialization.
}
