% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik_mat.R
\name{get_initialization_score}
\alias{get_initialization_score}
\title{Wrapper function to calculate the score for a single initialization.}
\usage{
get_initialization_score(X, y, beta_k, sigmaSq_k)
}
\arguments{
\item{beta_k}{posterior mean after a given initialization and moderate number of iterations.}

\item{sigmaSq_k}{estimated sigmaSq after a given initialization and moderate number of iterations.}
}
\value{
numeric, observed-data log-likelihood
}
\description{
Wrapper function to calculate the score for a single initialization.
}
