% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSamples_full.R
\name{getsamples.full}
\alias{getsamples.full}
\title{get posterior samples with a, b sampled from Metropolis-Hasting (full Bayes)}
\usage{
getsamples.full(
  num,
  X,
  y,
  a0,
  b0,
  beta0,
  nu0,
  lambda0,
  omega0,
  w0,
  sigmaSq0,
  omega_hyper,
  burn = 0,
  xi0 = NULL,
  tunea = 0.1,
  tuneb = 0.2,
  uppera = 1,
  upperb = 1,
  woodbury = FALSE,
  approx = FALSE,
  diagX = FALSE,
  prshapeo,
  prrateo,
  prshapes,
  prrates
)
}
\arguments{
\item{num}{numeric, number of samples to generate after burn-in for each hyperparameter}

\item{burn}{numeric, number of burn-in samples}

\item{xi0}{initial value for xi, the auxiliary variable, xi0 = NULL if not needed}

\item{tunea}{numeric from 0-1, step size of a in Metropolis-Hasting}

\item{tuneb}{numeric from 0-1, step size of b in Metropolis-Hasting}

\item{uppera}{upper bound for a, a ~ uniform(0, uppera)}

\item{upperb}{upper bound for b, b ~ uniform(0, upperb)}
}
\value{
a list of posterior samples generated, including beta, nu, phi, w, sigmaSq, a, b, xi, as well as acceptance rates of a and b
}
\description{
get posterior samples with a, b sampled from Metropolis-Hasting (full Bayes)
}
