% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConverged_emp.R
\name{isConverged_beta}
\alias{isConverged_beta}
\title{define another convergence criterion}
\usage{
isConverged_beta(beta_matrix, delta1, delta3, window_size = 3)
}
\arguments{
\item{beta_matrix}{A num_iteration * p matrix, where each row is the posterior mean estimate of beta at each iteration}

\item{window_size}{numeric, size of the sliding window}
}
\description{
define another convergence criterion
}
