% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConverged_emp.R
\name{isConverged_hyper}
\alias{isConverged_hyper}
\title{Define a convergence criterion. Several numeric vectors, each containing at least window_size + 1 elements.
The function calculates the relative change over the last window_size + 1 points for each vector individually, 
and checks whether all of them fall below the threshold delta2.}
\usage{
isConverged_hyper(..., delta1, delta2, window_size = 3)
}
\arguments{
\item{..., }{multiple numerical vectors}

\item{delta1, }{numeric, generally fix delta1 at 0.001}

\item{delta2, }{numeric}

\item{window_size}{numeric, size of the sliding window}
}
\description{
Define a convergence criterion. Several numeric vectors, each containing at least window_size + 1 elements.
The function calculates the relative change over the last window_size + 1 points for each vector individually, 
and checks whether all of them fall below the threshold delta2.
}
