% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isConverged_emp.R
\name{isConverged_rho}
\alias{isConverged_rho}
\title{convergence criterion based on the Hellinger distance of shrinkage factor densities}
\usage{
isConverged_rho(a_vec, b_vec, omega_vec, delta, window_size = 3)
}
\arguments{
\item{a_vec, }{vector of a}

\item{b_vec, }{vector of b}

\item{omega_vec, }{vector of omega}

\item{delta, }{numeric, error bound}

\item{window_size}{numeric, size of the sliding window}
}
\description{
convergence criterion based on the Hellinger distance of shrinkage factor densities
}
