% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MH.R
\name{mh}
\alias{mh}
\title{Metropolis-Hastings sampler, assuming an uniform prior unif(lower=0, upper)}
\usage{
mh(p, x0, vec, step, lower = 0, upper = 1)
}
\arguments{
\item{vec}{a vector that will be used, nu for sampling a and lambda for sampling b}

\item{step}{numeric, step size}
}
\value{
posterior samples and acceptance rates
}
\description{
Metropolis-Hastings sampler, assuming an uniform prior unif(lower=0, upper)
}
