% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_engine.R
\name{run_em_engine}
\alias{run_em_engine}
\title{The core EM-within-Gibbs engine.
This function runs the iterative EM algorithm until convergence or max_iter is reached.
It can handle fixed or updating hyperparameters based on the 'null.*' arguments.}
\usage{
run_em_engine(
  X,
  y,
  a_init,
  b_init,
  omega_init,
  sigmaSq_init,
  null.a,
  null.b,
  null.omega,
  null.sigmaSq,
  max_iter,
  iter_burnin,
  iter_samples,
  delta1 = 1e-06,
  delta2 = 0.001,
  delta3 = 0.001,
  window_size = 5,
  converge_rule = "hyper",
  woodbury = FALSE,
  approx = FALSE,
  IS_period = 1,
  option = "augmented",
  diagX = FALSE
)
}
\arguments{
\item{a_init, }{b_init, omega_init, sigmaSq_init Initial values for hyperparameters.}

\item{null.a, }{null.b, null.omega, null.sigmaSq Booleans indicating which parameters to update (TRUE) vs. keep fixed (FALSE).}

\item{max_iter}{Maximum number of iterations for this run.}

\item{iter_burnin}{number of burn-in samples in each iteration}

\item{iter_samples}{number of samples used for estimation in each iteration}

\item{window_size}{numeric, size of the sliding window}

\item{converge_rule}{"rho" or "hyper", specifying what to monitor to decide the convergence}

\item{IS_period}{The frequency of resampling vs. using IS}

\item{option}{"augmented" or "naive", specifying how we would like to sample nu}
}
\value{
A list containing the converged parameters (params) and the final state of the samplers (final_state).
}
\description{
The core EM-within-Gibbs engine.
This function runs the iterative EM algorithm until convergence or max_iter is reached.
It can handle fixed or updating hyperparameters based on the 'null.*' arguments.
}
