% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSamples_full_fixed_ab.R
\name{tpb_fullyBayes_fixab}
\alias{tpb_fullyBayes_fixab}
\title{Fully Bayesian Gibbs Sampler for the TPB Normal Model with Fixed a and b}
\usage{
tpb_fullyBayes_fixab(
  X,
  y,
  a,
  b,
  num_burnin,
  num_samples,
  woodbury = TRUE,
  diagX = FALSE
)
}
\arguments{
\item{num_burnin}{The number of MCMC iterations to discard as burn-in.}

\item{num_samples}{The number of MCMC iterations to save after the burn-in period.}
}
\value{
list of posterior samples
}
\description{
Fully Bayesian Gibbs Sampler for the TPB Normal Model with Fixed a and b
}
