% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_find_start.R
\name{initialize_adaptive}
\alias{initialize_adaptive}
\title{Performs the Data-Adaptive Initialization Strategy via Candidate Model Competition.}
\usage{
initialize_adaptive(
  X,
  y,
  omega_init_ebrr,
  sigmaSq_init_ebrr,
  iter_burnin,
  iter_samples,
  adapt_burnin,
  adapt_samples,
  adapt_iter,
  woodbury,
  IS_period = 1,
  delta1 = 1e-06,
  delta2 = 0.001,
  delta3 = 0.001,
  window_size = 5,
  converge_rule = "hyper",
  approx = FALSE,
  option = "augmented",
  diagX = FALSE
)
}
\arguments{
\item{X, }{y Model matrix and response vector.}

\item{adapt_burnin}{Burn-in iterations for the "clean" Gibbs run for scoring.}

\item{adapt_samples}{Sampling iterations for the "clean" Gibbs run for scoring.}

\item{adapt_iter}{Max number of iterations for the mini-EM of the adaptive initialization}

\item{omega_init_guess}{A single numeric value for the initial guess of omega.}

\item{sigmaSq_init_guess}{A single numeric value for the initial guess of sigmaSq.}

\item{...}{All other necessary parameters and functions to be passed to run_em_engine
(e.g., IS_period, converge_rule, getsamples.emp, log_lik, M.step functions, etc.).}
}
\value{
A list with the winning starting values: a.start, b.start, omega.start, sigmaSq.start
}
\description{
This function automates the selection of starting points for the main EM_GibbsTPB
algorithm by running a competition between several candidate models (special cases
of the TPB prior). It selects the candidate that best fits the data and returns
its parameters as the starting points.
}
